change payloadlen type size_t -> int

+++ librtp/include/rtp-packet.h
@@ -16,13 +16,13 @@ struct rtp_packet_t
 	uint16_t extlen; // extension length in bytes
 	uint16_t reserved; // extension reserved
 	const void* payload; // payload
-	size_t payloadlen; // payload length in bytes
+	int payloadlen; // payload length in bytes
 };
 
 ///@return 0-ok, other-error
-int rtp_packet_deserialize(struct rtp_packet_t *pkt, const void* data, size_t bytes);
+int rtp_packet_deserialize(struct rtp_packet_t *pkt, const void* data, int bytes);
 
 ///@return <0-error, >0-rtp packet size, =0-impossible
-int rtp_packet_serialize(const struct rtp_packet_t *pkt, void* data, size_t bytes);
+int rtp_packet_serialize(const struct rtp_packet_t *pkt, void* data, int bytes);
 
 #endif /* !_rtp_packet_h_ */
